Add files via upload
Arbitrary-precision arithmetic, also called big-num arithmetic, multiple precision arithmetic, or sometimes infinite-precision arithmetic, indicates that calculations are performed on numbers whose digits of precision are limited only by the available memory of the host system. This contrasts with the faster fixed-precision arithmetic found in most arithmetic logic unit (ALU) hardware, which typically offers between 8 and 64 bits of precision. 
A common application is public-key cryptography, whose algorithms commonly employ arithmetic with integers having hundreds of digits. The goal of this project is to implement basic mathematical, modulus and power-of operators of two large numbers.